Add Galaxy2 converted to Haskell
After the end of the ICFPC 2020 organizers left the following message:

> GALAXY v.2

> bugfixes?
> easter eggs?
> who knows...

This message has attachment "galaxy.v2.txt" with code.

Sources:
- attachment: https://discord.com/channels/728323267039068221/728323622757728417/734761715199115314
- message: https://discord.com/channels/728323267039068221/728323622757728417/734762487781785620

This is galaxy.v2.txt converted to Haskell by me (we tried to use
similar approach, but we didn't succeed to write sufficiently good
Prelude and we had other bugs).

This module compiles correctly, but I can't confirm that it works,
because I have no idea how to interact with galaxies.